+++ src/imio/smartweb/core/contents/pages/pages.py
@@ -1,6 +1,7 @@
 # -*- coding: utf-8 -*-
 
 from plone.dexterity.content import Container
+from zExceptions import BadRequest
 from zope.interface import Interface
 from zope.interface import implementer
 
@@ -16,6 +17,10 @@ class IDefaultPages(Interface):
 @implementer(IPages)
 class Pages(Container):
     """Shared base class for Pages"""
+
+    def checkValidId(self, id, allow_dup=0):
+        if hasattr(self, id) and id not in self.contentIds():
+            raise BadRequest()
 
     def canSetDefaultPage(self):
         return False